requirements.txt
django-allauth
django-heroku
django-storages
boto3
gunicorn
psycopg2